requirements.txt
pandas==2.1.1
numpy==1.26.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
minio==7.2.0
python-dotenv==1.0.0
requests==2.31.0
fastparquet==2023.10.1
pyarrow==14.0.1